This is a preparation for #59 where we want to eventually extend projected problems. This PR changes the storage of the `Proj_SPMF_NEP`. For memory reasons the projected NEP is created from the `pnep.projnep_B_mem` which is an vector of (large) matrices. The projected problems are created as SubArrays of this, when `set_proj_matrices!` is called.

This required changes in `nlar`, `tiar`, `jd`. 

Based on the benchmark report from travis, this PR creates a slight regression in nlar and jd, but I think we will take this back once we actually implement an `extend_proj_matrices!`: 
Before PR:
```
  Nonlinear Arnoldi                                          59.2 s    12.7 GiB
  Jacobi–Davidson                                            1.99 s     460 MiB
```
After PR:
```
  Nonlinear Arnoldi                                          64.5 s    13.2 GiB
  Jacobi–Davidson                                            2.02 s     467 MiB
```